backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
python-dotenv==1.0.1
pyotp==2.9.0
PyYAML==6.0.2
segno==1.6.1
gunicorn==22.0.0